requirements.txt
pyserial
bleak
pylgbst
influxdb_client
paho_mqtt
pygame